python_learning/gcp_python_projects/02_manage_gcp_storage_with_python/requirements.txt
google-cloud-storage==2.0.0
google-auth==2.0.0
